housekeeping

+++ 3-checks/2-reference.sh
@@ -1,8 +1,15 @@
 #!/usr/bin/env sh
 
 
-# Initialization
+# Lock file
+if [ -e lock ]; then
+	echo 'ERROR: Lock file found. Do not run multiple instances of this script.'
+	exit 1
+fi
+touch lock
 
+
+# Initialization
 platform=`../0-tools/platform.sh`
 hostname=`hostname`
 machine=$platform\_$hostname
@@ -13,9 +20,9 @@ if [ ! -d gathered ]; then
     exit 1
 fi
 
-if [ ! -e regression/$machine ]; then
-	mkdir -p regression/$machine
-fi
+#if [ ! -e regression/$machine ]; then
+#	mkdir -p regression/$machine
+#fi
 
 if [ ! -e blacklist ]; then
 	touch blacklist
@@ -25,6 +32,14 @@ if [ ! -e whitelist ]; then
 fi
 if [ -e log ]; then
 	truncate -s 0 log
+fi
+if [ -e errors ]; then
+	rm -f errors/*
+else
+	mkdir errors
+fi
+if [ ! -e builds ]; then
+	mkdir builds
 fi
 
 
@@ -43,57 +58,26 @@ cd ..
 
 # Fill worklist with example list
 if [ ! -s worklist ]; then
-	echo 'WARNING: No worklist with SHAs found, filling worklist with example SHAs'
+	echo 'WARNING: No worklist with SHAs found, using worklist with example SHAs'
 	cp -f worklist.example worklist
 fi
 
 
 # Iterating commit SHAs
 for sha in `cat worklist`; do
-	echo 'INFO: Testing commit '$sha
+	echo 'INFO: Verfying commit '$sha
 	if [ `grep -c $sha blacklist` -ne 0 ]; then
 		echo 'WARNING: Omitting blacklisted commit '$sha
 		continue
 	fi
-	cd nuspell
+	# only build when previously build executable is not available
+	if [ ! -e builds/$sha ]; then
+		cd nuspell
 
-	# Reset to the desired commit SHA
-	git reset --hard $sha >> ../log 2>> ../log
-	if [ $? -ne 0 ]; then
-		echo 'ERROR: Resetting to commit '$sha' results in an error'
-		cd ..
-		# Add commit SHA to blacklist
-		echo $sha >> blacklist
-		cat blacklist|sort|uniq > blacklist.tmp
-		mv -f blacklist.tmp blacklist
-		continue
-	fi
-
-	# Build application
-	autoreconf -vfi >> ../log 2>> ../log
-	if [ $? -ne 0 ]; then
-		echo 'ERROR: Failed to automatically reconfigure nuspell'
-		cd ..
-		# Add commit SHA to blacklist
-		echo $sha >> blacklist
-		cat blacklist|sort|uniq > blacklist.tmp
-		mv -f blacklist.tmp blacklist
-		continue
-	fi
-	./configure CXXFLAGS='-O2 -fno-omit-frame-pointer' >> ../log 2>> ../log
-	if [ $? -ne 0 ]; then
-		echo 'ERROR: Failed to configure nuspell'
-		cd ..
-		# Add commit SHA to blacklist
-		echo $sha >> blacklist
-		cat blacklist|sort|uniq > blacklist.tmp
-		mv -f blacklist.tmp blacklist
-		continue
-	fi
-	if [ `grep -c $sha ../whitelist` -eq 0 ]; then
-		make -j CXXFLAGS='-O2 -fno-omit-frame-pointer' >> ../log 2>> ../log
+		# Reset to the desired commit SHA
+		git reset --hard $sha >> ../log 2>> ../log
 		if [ $? -ne 0 ]; then
-			echo 'ERROR: Failed to build nuspell'
+			echo 'ERROR: Resetting to commit '$sha' results in an error'
 			cd ..
 			# Add commit SHA to blacklist
 			echo $sha >> blacklist
@@ -101,9 +85,11 @@ for sha in `cat worklist`; do
 			mv -f blacklist.tmp blacklist
 			continue
 		fi
-		make check >> ../log 2>> ../log
+
+		# Build application
+		autoreconf -vfi >> ../log 2>> ../log
 		if [ $? -ne 0 ]; then
-			echo 'ERROR: Failed to check nuspell'
+			echo 'ERROR: Failed to automatically reconfigure nuspell'
 			cd ..
 			# Add commit SHA to blacklist
 			echo $sha >> blacklist
@@ -111,13 +97,51 @@ for sha in `cat worklist`; do
 			mv -f blacklist.tmp blacklist
 			continue
 		fi
-	else
-		# Faster building for whitelisted commit
-		cd src/hunspell
-		make -j CXXFLAGS='-O2 -fno-omit-frame-pointer' libhunspell.a >> ../../../log 2>> ../../../log
-		cd ../nuspell
-		make -j CXXFLAGS='-O2 -fno-omit-frame-pointer' regress >> ../../../log 2>> ../../../log
-		cd ../..
+		./configure CXXFLAGS='-O2 -fno-omit-frame-pointer' CPPFLAGS='-DHUNSPELL_WARNING_ON' >> ../log 2>> ../log
+		if [ $? -ne 0 ]; then
+			echo 'ERROR: Failed to configure nuspell'
+			cd ..
+			# Add commit SHA to blacklist
+			echo $sha >> blacklist
+			cat blacklist|sort|uniq > blacklist.tmp
+			mv -f blacklist.tmp blacklist
+			continue
+		fi
+		if [ `grep -c $sha ../whitelist` -eq 0 ]; then
+			make -j >> ../log 2>> ../log
+			if [ $? -ne 0 ]; then
+				echo 'ERROR: Failed to build nuspell'
+				cd ..
+				# Add commit SHA to blacklist
+				echo $sha >> blacklist
+				cat blacklist|sort|uniq > blacklist.tmp
+				mv -f blacklist.tmp blacklist
+				continue
+			fi
+			make -j check >> ../log 2>> ../log
+			if [ $? -ne 0 ]; then
+				echo 'ERROR: Failed to check nuspell'
+				cd ..
+				# Add commit SHA to blacklist
+				echo $sha >> blacklist
+				cat blacklist|sort|uniq > blacklist.tmp
+				mv -f blacklist.tmp blacklist
+				continue
+			fi
+		else
+			# Faster building for whitelisted commit
+			cd src/hunspell
+			make -j libhunspell.a >> ../../../log 2>> ../../../log
+			cd ../nuspell
+			if [ -e verify.cxx ]; then
+				make -j verify >> ../../../log 2>> ../../../log
+				cp verify ../../../builds/$sha
+			else
+				make -j regress >> ../../../log 2>> ../../../log
+				cp regress ../../../builds/$sha
+			fi
+			cd ../..
+		fi
 	fi
 
 	timestamp=`git log $sha --date=raw|head -n 3|tail -n 1|awk '{print $2}'`
@@ -133,12 +157,13 @@ for sha in `cat worklist`; do
 		if [ -e gathered/$language/words ]; then
 			wordsin=`cat gathered/$language/words.total`
 			echo 'Testing '$language' on '$wordsin' words'
-			mkdir -p regression/$machine/$language
+#			mkdir -p regression/$machine/$language
 			# Add commit SHA, commit timestamp and run timestamp to result
-			echo -n $sha' '$timestamp' '$run' ' >> regression/$machine/$language/result
-			nuspell/src/nuspell/regress -i UTF-8 -d $dictionary gathered/$language/words 2> regression/$machine/$language/stderr |tail -n 1 >> regression/$machine/$language/result
+			echo -n $sha' '$timestamp' '$language' '$run' ' >> $machine.tsv
+			result=`builds/$sha -i UTF-8 -d $dictionary gathered/$language/words 2> errors/$language |tail -n 1`
+			echo $result >> $machine.tsv
 			# Double check if number of words in word list and number of words tested are identical
-			wordsout=`tail -n 1 regression/$machine/$language/result | awk '{print $4}'`
+			wordsout=`echo $result | awk '{print $1}'`
 			if [ $wordsin -ne $wordsout ]; then
 				echo 'ERROR: Number of words in ('$wordsin') and number of results out ('$wordsout') do not match for '$language
 			fi
@@ -155,3 +180,7 @@ done
 
 # Clear worklist
 truncate -s 0 worklist
+
+
+# Lock file
+rm -f lock
